rename encryptionkey to encryptionsearcher resource
