+++ service/controller/control_plane_resource_set.go
@@ -28,7 +28,7 @@ import (
 	"github.com/giantswarm/aws-operator/service/controller/resource/accountid"
 	"github.com/giantswarm/aws-operator/service/controller/resource/awsclient"
 	"github.com/giantswarm/aws-operator/service/controller/resource/cpvpc"
-	"github.com/giantswarm/aws-operator/service/controller/resource/encryptionkey"
+	"github.com/giantswarm/aws-operator/service/controller/resource/encryptionsearcher"
 	"github.com/giantswarm/aws-operator/service/controller/resource/region"
 	"github.com/giantswarm/aws-operator/service/controller/resource/s3object"
 	"github.com/giantswarm/aws-operator/service/controller/resource/snapshotid"
@@ -193,16 +193,16 @@ func newControlPlaneResourceSet(config controlPlaneResourceSetConfig) (*controll
 		}
 	}
 
-	var encryptionKeyResource resource.Interface
+	var encryptionSearcherResource resource.Interface
 	{
-		c := encryptionkey.Config{
+		c := encryptionsearcher.Config{
 			G8sClient:     config.G8sClient,
 			Encrypter:     encrypterObject,
 			Logger:        config.Logger,
 			ToClusterFunc: newControlPlaneToClusterFunc(config.G8sClient),
 		}
 
-		encryptionKeyResource, err = encryptionkey.New(c)
+		encryptionSearcherResource, err = encryptionsearcher.New(c)
 		if err != nil {
 			return nil, microerror.Mask(err)
 		}
@@ -347,7 +347,7 @@ func newControlPlaneResourceSet(config controlPlaneResourceSetConfig) (*controll
 		// into the controller context.
 		awsClientResource,
 		accountIDResource,
-		encryptionKeyResource,
+		encryptionSearcherResource,
 		tccpnOutputsResource,
 		snapshotIDResource,
 		tccpAZsResource,
